PDMain/requirements.txt
discord.py>=2.3.2,<3.0
aiohttp>=3.9.0,<4.0
python-dotenv>=1.0.0,<2.0
aiosqlite>=0.19.0,<1.0
pytest>=7.0.0,<8.0
pytest-asyncio>=0.21.0,<1.0
